+++ qiskit/ignis/verification/topological_codes/circuits.py
@@ -12,28 +12,26 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-'''
-Generates circuits for quantum error correction
-'''
+"""Generates circuits for quantum error correction."""
 
 from qiskit import QuantumRegister, ClassicalRegister
 from qiskit import QuantumCircuit
 
 
 class RepetitionCode():
-    '''
+    """
     Implementation of a distance d repetition code, implemented over
     T syndrome measurement rounds.
-    '''
+    """
 
     def __init__(self, d, T=0):
-        '''
+        """
         Creates the circuits corresponding to a logical 0 and 1 encoded
         using a repetition code.
 
         Args:
             d: Number of code qubits (and hence repetitions) used.
-            T: Number of rounds of ancilla-assited syndrome measurement.
+            T: Number of rounds of ancilla-assisted syndrome measurement.
 
 
         Additional information:
@@ -41,7 +39,7 @@ class RepetitionCode():
             `T` rounds are added, followed by measurement of the code
             qubits (corresponding to a logical measurement and final
             syndrome measurement round).
-        '''
+        """
 
         self.d = d
         self.T = 0
@@ -67,38 +65,36 @@ class RepetitionCode():
             self.readout()
 
     def get_circuit_list(self):
-        '''
+        """
         Returns:
             circuit_list: self.circuit as a list, with
             circuit_list[0] = circuit['0']
             circuit_list[1] = circuit['1']
-        '''
+        """
         circuit_list = [self.circuit[log] for log in ['0', '1']]
         return circuit_list
 
     def x(self, logs=('0', '1')):
-        '''
+        """
         Applies a logical x to the circuits for the given logical values.
 
         Args:
             logs: List or tuple of logical values expressed as strings.
-        '''
+        """
         for log in logs:
             for j in range(self.d):
                 self.circuit[log].x(self.code_qubit[j])
             self.circuit[log].barrier()
 
     def _preparation(self):
-        '''
+        """
         Prepares logical bit states by applying an x to the circuit that will
         encode a 1.
-        '''
+        """
         self.x(['1'])
 
     def syndrome_measurement(self):
-        '''
-        Application of a syndrome measurement round.
-        '''
+        """Application of a syndrome measurement round."""
 
         self.link_bits.append(ClassicalRegister(
             (self.d - 1), 'round_' + str(self.T) + '_link_bit'))
@@ -124,16 +120,16 @@ class RepetitionCode():
         self.T += 1
 
     def readout(self):
-        '''
+        """
         Readout of all code qubits, which corresponds to a logical measurement
         as well as allowing for a measurement of the syndrome to be inferred.
-        '''
+        """
         for log in ['0', '1']:
             self.circuit[log].add_register(self.code_bit)
             self.circuit[log].measure(self.code_qubit, self.code_bit)
 
     def process_results(self, raw_results):
-        '''
+        """
         Args:
             raw_results: A dictionary whose keys are logical values, and whose
             values are standard counts dictionaries, (as obtained from the
@@ -149,7 +145,7 @@ class RepetitionCode():
             their is full freedom to compile, choose a backend, use a
             noise model, etc. The results from these executions should then
             be used to create the input for this method.
-        '''
+        """
         results = {}
         for log in raw_results:
             results[log] = {}